// backend/src/types/EDataSourceType.ts
export enum EDataSourceType {
    EXCEL = 'excel',
    CSV = 'csv',
    PDF = 'pdf',
    POSTGRESQL = 'postgresql',
    MYSQL = 'mysql',
    MARIADB = 'mariadb',
    MONGODB = 'mongodb',
    GOOGLE_ANALYTICS = 'google_analytics',
    GOOGLE_AD_MANAGER = 'google_ad_manager',
    GOOGLE_ADS = 'google_ads',
    META_ADS = 'meta_ads',
}